+++ Cargo.toml
@@ -316,8 +316,8 @@ vortex-fsst = { version = "0.1.0", path = "./encodings/fsst", default-features =
 vortex-io = { version = "0.1.0", path = "./vortex-io", default-features = false }
 vortex-ipc = { version = "0.1.0", path = "./vortex-ipc", default-features = false }
 vortex-json = { version = "0.1.0", path = "./vortex-json", default-features = false }
-vortex-layout-commons = { version = "0.1.0", path = "./vortex-layout-commons", default-features = false }
 vortex-layout = { version = "0.1.0", path = "./vortex-layout", default-features = false }
+vortex-layout-commons = { version = "0.1.0", path = "./vortex-layout-commons", default-features = false }
 vortex-mask = { version = "0.1.0", path = "./vortex-mask", default-features = false }
 vortex-metrics = { version = "0.1.0", path = "./vortex-metrics", default-features = false }
 vortex-onpair = { version = "0.1.0", path = "./encodings/onpair", default-features = false }
