+++ vortex-layout-commons/Cargo.toml
@@ -19,33 +19,23 @@ all-features = true
 [dependencies]
 arcref = { workspace = true }
 async-trait = { workspace = true }
-futures = { workspace = true, features = ["alloc", "async-await"] }
 flatbuffers = { workspace = true }
+futures = { workspace = true, features = ["alloc", "async-await"] }
 itertools = { workspace = true }
-moka = { workspace = true, features = ["future"] }
 once_cell = { workspace = true, features = ["parking_lot"] }
 parking_lot = { workspace = true }
 pin-project-lite = { workspace = true }
-rustc-hash = { workspace = true }
 termtree = { workspace = true }
-tokio = { workspace = true, features = ["rt"], optional = true }
-tracing = { workspace = true }
 vortex-array = { workspace = true }
 vortex-buffer = { workspace = true }
 vortex-error = { workspace = true }
 vortex-flatbuffers = { workspace = true, features = ["layout"] }
 vortex-mask = { workspace = true }
-vortex-metrics = { workspace = true }
 vortex-session = { workspace = true }
-vortex-utils = { workspace = true, features = ["dashmap"] }
+vortex-utils = { workspace = true }
 
 [dev-dependencies]
 rstest = { workspace = true }
-tokio = { workspace = true, features = ["rt", "macros"] }
-
-[features]
-_test-harness = []
-tokio = ["dep:tokio", "vortex-error/tokio"]
 
 [lints]
 workspace = true
