+++ vortex-layout/src/segments/cache.rs
