+++ vortex-layout/src/segments/shared.rs
